+++ database/migrations/2018_12_13_141119_create_comments_table.php
@@ -13,10 +13,6 @@ class CreateCommentsTable extends Migration
      */
     public function up()
     {
-        if (!env('APP_DEBUG')) {
-            echo "Not In Test Environment! \n";
-            return;
-        }
         if (Schema::hasTable('comments')) {
             echo "Table comments Is Already Exist! \n";
             return;
@@ -25,15 +21,33 @@ class CreateCommentsTable extends Migration
         Schema::create('comments', function (Blueprint $table) {
             $table->charset = 'utf8mb4';
             $table->increments('id');
-            $table->text('main_body'); // 内容
+            $table->string('body', 64); // 内容
             $table->unsignedInteger('user_id'); // table:users->id
             $table->unsignedInteger('article_id'); // table:articles->id
-            $table->unsignedTinyInteger('examine')->default(0); // 审核状态:0=未触发,1=待审核,2=通过,3=拒绝
+            $table->unsignedTinyInteger('examine')->default(0);
+            // DB:comments->examine(审核状态):0=未触发,1=待审核,2=通过,3=拒绝
             $table->boolean('blacklist')->default(false); // 黑名单
             $table->softDeletes();
             $table->timestamps();
+            $table->index('user_id');
+            $table->index('article_id');
         });
         // \Log::debug(\DB::getQueryLog());
+
+        // create table `comments` (
+        // `id` int unsigned not null auto_increment primary key,
+        // `body` varchar(64) not null,
+        // `user_id` int unsigned not null,
+        // `article_id` int unsigned not null,
+        // `examine` tinyint unsigned not null default '0',
+        // `blacklist` tinyint(1) not null default '0',
+        // `deleted_at` timestamp null,
+        // `created_at` timestamp default CURRENT_TIMESTAMP null,
+        // `updated_at` timestamp ON UPDATE CURRENT_TIMESTAMP null
+        // ) default character set utf8mb4 collate 'utf8mb4_unicode_ci'
+
+        // alter table `comments` add index `comments_user_id_index`(`user_id`)
+        // alter table `comments` add index `comments_article_id_index`(`article_id`)
     }
 
     /**
@@ -43,6 +57,10 @@ class CreateCommentsTable extends Migration
      */
     public function down()
     {
+        if (!env('APP_DEBUG')) {
+            echo "Not In Test Environment! \n";
+            return;
+        }
         Schema::dropIfExists('comments');
     }
 }
